+++ packages/zowe-explorer/CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to the "vscode-extension-for-zowe" extension will be documen
 - Updated linter rules and addressed linter errors throughout the codebase. [#2184](https://github.com/zowe/vscode-extension-for-zowe/issues/2184)
 - Added a new setting `zowe.files.logsFolder.path` that can be used to override Zowe Explorer logs folder if default location is read-only. [#2186](https://github.com/zowe/vscode-extension-for-zowe/issues/2186)
 - Spool files will now refresh if the file is open in the editor and clicked again in the tree view. [#1952](https://github.com/zowe/vscode-extension-for-zowe/issues/1952)
+- Added a new context menu option to allow users to download Job Spool content in binary format. [#2060](https://github.com/zowe/vscode-extension-for-zowe/issues/2060)
 
 ### Bug fixes